requirements.txt
aiogram==3.7.0
aiosqlite==0.20.0
openpyxl==3.1.5
reportlab==4.1.0
python-dotenv==1.0.1
